+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Anagrams".
 - Tests for "The stonemason".
 - Tests for "Continued fractions".
+- Tests for "All operations are equal!".
 
 ### Fixed
 - Covers annotation for "The grand festival - II".